refactor payment service and move notifications
