+++ frontend/src/services/notificationService.js
@@ -26,3 +26,29 @@ export const createNotification = async (payload) => {
   const res = await api.post('/notifications', payload);
   return res.data.data || res.data;
 };
+
+// Local notification utilities
+const notifications = [];
+
+const notificationService = {
+  push: (type, message, data = {}) => {
+    const id = Date.now().toString();
+    notifications.push({ id, type, message, data, read: false, timestamp: new Date() });
+    console.log(`[NOTIFY] ${type}: ${message}`);
+  },
+
+  getAll: async () => {
+    return notifications;
+  },
+
+  markAsRead: async (id) => {
+    const notif = notifications.find((n) => n.id === id);
+    if (notif) notif.read = true;
+  },
+
+  clearAll: async () => {
+    notifications.length = 0;
+  },
+};
+
+export default notificationService;
